Fix min() type mismatch for Arduino ESP32 3.x / GCC 13
Preferences::getInt() returns int32_t in Arduino 3.x; std::min requires
matching types unlike the Arduino 2.x macro. Cast to int to match kWebHistoryMax.

+++ src/AircraftPersistence.cpp
@@ -43,7 +43,7 @@ void AircraftPersistence::saveCounts(const int* counts) const {
 int AircraftPersistence::loadHistory(Aircraft* history) const {
     Preferences prefs;
     prefs.begin(kNvsNamespace, true);
-    int count = min(prefs.getInt("histCount", 0), kWebHistoryMax);
+    int count = min((int)prefs.getInt("histCount", 0), kWebHistoryMax);
     if (count > 0) {
         StoredAircraft buf[kWebHistoryMax] = {};
         prefs.getBytes("histData", buf, sizeof(StoredAircraft) * count);
